// frontend/src/model/Student.ts
interface Student {
    iden: string;
    name: string;
    nim: string;
    semester: string;
    tempat_magang: string;
    email: string,
    phone: string
    image_url: string;
    status: string;
    faculty_supervisor: string;
    site_supervisor: string;
    major: string;
    notes: string;
    period: string;
}

export default Student;